Remove rbd write-cache docs
After some usage of rbd write feature, we came to
the conclusion that it is not mature yet to work
with openstack.

The issue we currently see is when a vm with write-cache
enabled is rebooted, it doesnt come bad and error with
https://paste.opendev.org/raw/817888/

We dont recomment the enablement of that feature for now.