fix: tune net.core.netdev_budget
The default net.core.netdev_budget for the kernel is 300 which is
not enough for most modern systems which have 10G networking, which
is the common case in any OpenStack node.  This patch tunes that up
to 1000 which is a much better number for that.

This patch also removes the `NodeSoftNetTimesSqueezed` alert since
it is not actionable and it is something that can normally happen
as a function of a system that has very high traffic.  The bigger
concern is the NodeSoftNetDrops which means that there was so much
packets in the queue that were not processed in time that packets
were dropped.

If things take a little longer, there's not that much we can do
about it.