fix(nova): set list_records_by_skipping_down_cells to false
If a cell is down, OpenStack responds with a 200 OK but the instances
missing when using the older micro-versions.  This is very confusing
for users and most of the time there is only a single cell anyways
that is in place.

It makes sense to just give a 500 error instead, the user will
work much better.