Keep glance-images pvc when change glance storage
Glance image PVC contains requires images that should be careful handled
during changing glance storage when migrate out of PVC mode for glance
storage.
Which in migrate/upgrade path, should be be correctly moved and deleted
the PVC after.

On the other hand, it's also possible to accidentally changes storage
mode out of `pvc` and lose the `glance-images` PVC which is unbearable
mistake.

Once storage mode set to `pvc, we should allow that PVC to be able to
stay and ready for reuse again until it's mannually deleted.