fix(uwsgi): tune settings
The current uWSGI settings that we ship are missing a few important
ones, specifically `http-raw-body` which is used by Glance in order
to enable chunked uploads.

This is the reason why a direct upload to Glance works just fine
because it goes through the ingress (which does the chunking for
us), however, it fails when creating an image from volume due to
the fact that Cinder talks directly to Glance without the ingress
being in the path.

Closes: #1046

Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>