Add rebuild time for images
If we want to rebuild all images for security reasons or when
branching a new version of OpenStack, we need to be able to
rebuild all images.  Doing this change in the job definition
is not enough to cause the promotion to happen.

This adds a tool that can modify the file by passing a Dockerfile
or many to it.