[ATMOSPHERE-514] fix ovn for backward compatibility (#2042)
we used to keep type: hv|gw labels in ovn-controller and ovn-controller-gw daemonsents to distinguish them.
in the PR #2023, we removed ovn-controller-gw daemonset and keep only one DS to sync with upstream as much as possible. so type: xx label is unnecessary. But we cannot remove it for running clouds because spec.selector field is immutable.
this pr reverts that label for backward compatibility.

Reviewed-by: Mohammed Naser <mnaser@vexxhost.com>