fix openvswitch directory permission issue when using dkdk
By default ovn and openvswitch run with root priviledge, when dpdk is
enabled, a vhost-user-client type port is created in OVS br-int,
OVS acts as the client and QEMU the server. nova need to create
vhostuser socket file in openvswitch run dir which used by QEMU.
therefore the openvswitch run dir need to be accessable by the
nova uid 42424.

Change-Id: Ib249e15240e9feb1c1029cba4b5bcc4f018fbc10