Add Rust based ovsinit binary
This binary allows us to more reliably migrate IPs, which also
includes routes as well, with proper rollback in place to avoid
a system falling off the network.

This first iteration is used in the OVN chart only, but the long
term plan is to leverage it into the Neutron and Open vSwitch
charts potentially to have a single "auto_bridge_add" source
of truth.

Change-Id: Ic4de23297b67a602d9aba4b00f0fb234d9d37cfe