Stop running ethtool during deployments
Running ethtool on the fly is now disabled by default.
This is to prevent network interruptions when the ethtool
command is run. The ethtool command is run in a separate
script that is executed by networkd-dispatcher when the
interface is up, therefore, on boot as well.

This patch set ethtool_apply variable to false by default.

Signed-off-by: Guilherme Steinmuller <gsteinmuller@vexxhost.com>